Split agent harness tests into focused modules (#603).
Moves shared fixtures into harness_shared.py and splits the monolithic
test_examples.py into smoke, MCP, LangGraph, and adapter suites so
contributors can navigate and extend harness coverage incrementally.
