+++ examples/agents/tests/conftest.py
@@ -1,0 +1,12 @@
+"""Pytest plugin surface for agent harness tests."""
+
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+_AGENTS_DIR = Path(__file__).resolve().parents[1]
+if str(_AGENTS_DIR) not in sys.path:
+    sys.path.insert(0, str(_AGENTS_DIR))
+
+from harness_shared import *  # noqa: F403
