+++ pyproject.toml
@@ -90,6 +90,10 @@ target-version = "py311"
 select = ["E", "F", "I"]
 ignore = ["E501"]
 
+[tool.ruff.lint.per-file-ignores]
+"examples/agents/tests/conftest.py" = ["E402"]
+"examples/agents/tests/harness/*.py" = ["E402"]
+
 [tool.ruff.format]
 quote-style = "double"
